shoutourbiz/requirements.txt
amqp==2.1.4
beautifulsoup4==4.6.0
billiard==3.5.0.2
celery==4.0.2
Django==1.8
django-countries==4.5
django-debug-toolbar==1.8
django-haystack==2.6.1
django-widget-tweaks==1.4.1
elasticsearch==5.4.0
httplib2==0.9.2
kombu==4.0.2
lxml==3.8.0
numpy==1.13.0
oauthlib==2.0.1
pandas==0.20.2
Pillow==3.4.2
python-dateutil==2.6.0
python-instagram==1.3.2
pytrends==4.1.1
pytz==2016.10
redis==2.10.5
reportlab==3.3.0
requests==2.12.3
requests-oauthlib==0.7.0
simplejson==3.10.0
six==1.10.0
sqlparse==0.2.2
tweepy==3.5.0
vine==1.1.3
stripe==1.62.0
django-forms-bootstrap==3.1.0
django-stripe-payments==2.0.0
django-bootstrap-form==2.0.5